121. Best Time to Buy and Sell Stock
Intuition:

The goal is to find the maximum profit from buying and selling a stock, where buying occurs at an earlier day and selling at a later day. We need to track the highest possible difference between prices in a way that ensures we always buy before we sell.

Approach:

We iterate through the prices array using two pointers:

- left to represent the buying day.
- right to represent the selling day.
- If the price at right is higher than the price at left, we calculate the profit (prices[right] - prices[left]).
- If the profit is higher than the current MaxProfit, we update MaxProfit.
- If the price at right is lower than the price at left, we set left to right because it's better to "buy" at the new, lower price.

Complexity
Time complexity: O(n)
We iterate through the prices list once, so the time complexity is linear.

Space complexity: O(1)
We are using only a constant amount of extra space (for the variables left, right, and MaxProfit).